Updated course domain layer
