+++ TODO
@@ -13,4 +13,8 @@
 ✔ Increase Jest Test times for some tests that potentially can run longer than 5 seconds @high @done(20-11-29 20:14)
 ✔ Refactor code to be cleaner - Clean engineering practises @done(20-11-30 08:41)
 ✔ Refactor bad name createOrUpdate to createOrUpdateCourse @high @done(20-12-03 15:39)
-☐ Error Handling very primitive - https://www.joyent.com/node-js/production/design/errors
+✔ Error Handling very primitive - https://www.joyent.com/node-js/production/design/errors @critical @done(20-12-05 14:54)
+☐ Refactor error handling to Middleware
+☐ Refactor use of index.js making it hard to debug or find stuff @high
+    ✔ Refactor course domain @done(20-12-05 14:56)
+    ☐ Refactor anything that has code in the index.js
